// src/app/system/resources/resource.ts
export class Resource {
  id: number;
  createdBy: number;
  createDate: number;
  lastModifiedBy: number;
  lastModifiedDate: number;
  parentId: number;
  lft: number;
  rt: number;
  level: number;
  type: number;
  name: string;
  permission: string;
  pathExp: string;
  method: string;
  icon: string;
  status: number;
}